refactor(agent): extract startup dispatch orchestration (#224)
